pyproject.toml
[tool.poetry]
name = "minesweeper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
pygame-minesweeper-core = "^1.0.18"
Flask-SocketIO = "^5.0.1"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
